Lua #41 - basic scripting support (#44)
First iteration of scripting in lua. There is a new folder 'scripts' in the repository, master.lua should not be changed, the other lua files contain dynamic values for each of the entities. Making changes there and restarting the game or pressing F5 in the game should propagate these values.